chore: update all.import-multiple
