+++ tests/spec/test-data/all.import-multiple/with-imports-for-create.component.ts
@@ -1,0 +1,11 @@
+import { ADep, BDep } from '../my/relative/path';
+import { Router, DDep } from '@angular/router';
+
+export class WithImportsComponent {
+    constructor(
+        private aDep: ADep,
+        private d: DDep,
+        private b: BDep,
+        private router: Router,
+    ) {}
+}
